refactor: Change deposit points and weight data type to float
Updates the data type for deposit points and weight from integer to float across the navigation graph and related screens.

- In `AuthenticatedNavHost`, the `navArgument` types for the `SuccessDeposit` route are changed from `NavType.IntType` to `NavType.FloatType`.
- In `SuccessDepositScreen`, the `points` and `weight` parameters are updated from `Int` to `Float` to correctly handle the floating-point values passed through navigation.

+++ app/src/main/java/com/ahargunyllib/growth/presentation/ui/navigation/nav_obj/AuthenticatedNavObj.kt
@@ -8,7 +8,7 @@ sealed class AuthenticatedNavObj(val route: String) {
     data object ExchangePoint: AuthenticatedNavObj("exchange_point")
     data object ScanQR: AuthenticatedNavObj("scan_qr")
     data object SuccessDeposit : AuthenticatedNavObj("success_deposit_screen/{points}/{weight}") {
-        fun createRoute(points: Float, weight: Float) = "success_deposit_screen/$points/$weight"
+        fun createRoute(points: Int, weight: Float) = "success_deposit_screen/$points/$weight"
     }
     
     data object AchievementScreen: AuthenticatedNavObj("achievement")
